# DecodeVarint: Fix panic on truncated input; Add tests