--- a/tests/molecule_test.go
+++ b/tests/molecule_test.go
@@ -126,3 +126,70 @@ func TestMoleculeSimple(t *testing.T) {
 		require.NoError(t, err)
 	}
 }
+
+type FuzzSplitBytes struct {
+	bytes    []byte
+	selected int
+}
+
+func (f *FuzzSplitBytes) Fuzz(c fuzz.Continue) {
+	f.selected = c.Intn(len(f.bytes))
+}
+
+func fuzzSplitBytes(fuzzer *fuzz.Fuzzer, bytes []byte) []byte {
+	splitter := &FuzzSplitBytes{bytes, 0}
+	fuzzer.Fuzz(splitter)
+	return bytes[:splitter.selected]
+}
+
+func TestMoleculeTruncatedShouldNotPanic(t *testing.T) {
+	var (
+		seed      = time.Now().UnixNano()
+		fuzzer    = fuzz.NewWithSeed(seed)
+		numFuzzes = 100000
+	)
+	defer func() {
+		// Log the seed to make debugging failures easier.
+		t.Logf("Running test with seed: %d", seed)
+	}()
+	// Limit slice size to prevent tests from taking too long.
+	fuzzer.NumElements(0, 100)
+
+	for i := 0; i < numFuzzes; i++ {
+		m := &simple.Simple{}
+		fuzzer.Fuzz(&m)
+		if m == nil {
+			continue
+		}
+
+		marshaled, err := proto.Marshal(m)
+		require.NoError(t, err)
+		require.NotEmpty(t, marshaled)
+
+		// split the buffer: iterating over it should not panic
+		// there was a bug in DecodeVarint that caused panics
+		splitMarshaled := fuzzSplitBytes(fuzzer, marshaled)
+		buffer := codec.NewBuffer(splitMarshaled)
+		err = molecule.MessageEach(buffer, func(fieldNum int32, value molecule.Value) (bool, error) {
+			return true, nil
+		})
+		// sometimes this split will actually generate a "correct" truncation: err == nil is okay
+		if err != nil {
+			require.Error(t, err, "unexpected EOF", "wtf %#v %#v", string(splitMarshaled), m)
+		}
+	}
+}
+
+func TestMoleculeTruncated(t *testing.T) {
+	payload := &simple.Simple{Int64: 42}
+	serialized, err := proto.Marshal(payload)
+	require.NoError(t, err)
+
+	// truncate the payload at the 1 byte mark: parsing should return an error, not panic
+	// this is a simplified case of a bug found by the TestMoleculeTruncatedShouldNotPanic fuzzer
+	buffer := codec.NewBuffer(serialized[:1])
+	err = molecule.MessageEach(buffer, func(fieldNum int32, value molecule.Value) (bool, error) {
+		return true, nil
+	})
+	require.Error(t, err, "unexpected EOF")
+}