@kevinconaway 2021-11-08
Will you please help me understand how the length is changing?