@evanj 2021-11-10
The length isn't changing. It is just that the existing code only checks `cb.len`, while the remaining space in the buffer is actually `cb.len - cb.index` (which is what gets returned by `cb.Len()`). The way the buffer gets used is the *entire* protocol buffer message is put in in, then `cb.index` is advanced towards the end.

For *correctly* encoded data, it doesn't matter: the current encoder will never "run off the end". For example, even in the "buffer has > 10 space" branch, if the Varint is terminated, it will return early: https://github.com/richardartoul/molecule/blob/master/src/codec/decode.go#L59-L62

The problem is when the varint is *incorrectly* encoded, this causes it to run off the end of the buffer.

As a simple example, consider the protocol buffer message: 

```
payload := &simple.Simple{Int64: 42}
```

This encodes as: `0x20 0x2a`. This gets decoded with two calls to DecodeVarint:

```
DecodeVarint index=0 len=2
DecodeVarint index=1 len=2
```

If you modify the payload to be `0x20 0xff` then it runs off the end:

```
DecodeVarint index=0 len=2
DecodeVarint index=1 len=2
panic: runtime error: index out of range [2] with length 2 [recovered]

goroutine 35 [running]:
testing.tRunner.func1.2({0x13812e0, 0xc00014e168})
	/usr/local/Cellar/go/1.17.2/libexec/src/testing/testing.go:1209 +0x24e
testing.tRunner.func1()
	/usr/local/Cellar/go/1.17.2/libexec/src/testing/testing.go:1212 +0x218
panic({0x13812e0, 0xc00014e168})
	/usr/local/Cellar/go/1.17.2/libexec/src/runtime/panic.go:1038 +0x215
github.com/richardartoul/molecule/src/codec.(*Buffer).DecodeVarint(0xc000133f10)
	/Users/evan.jones/molecule/src/codec/decode.go:170 +0x9bd
github.com/richardartoul/molecule.MessageEach(0xc000133f10, 0x13bb4a8)
	/Users/evan.jones/molecule/molecule.go:32 +0x105
github.com/richardartoul/molecule/tests.TestMessageEach(0xc00020a1a0)
	/Users/evan.jones/molecule/tests/molecule_test.go:145 +0x28f
testing.tRunner(0xc00020a1a0, 0x13bb4b0)
	/usr/local/Cellar/go/1.17.2/libexec/src/testing/testing.go:1259 +0x102
created by testing.(*T).Run
	/usr/local/Cellar/go/1.17.2/libexec/src/testing/testing.go:1306 +0x35a
FAIL	github.com/richardartoul/molecule/tests	0.719s
```

This happens because the code assumed the buffer had 2 bytes in it, from the following code, but it actually only had 1:

https://github.com/richardartoul/molecule/blob/master/src/codec/decode.go#L162-L164