@kevinconaway 2021-11-11
I understand, thanks.  Would you please post a benchstat diff?