@evanj 2021-11-12
I ran `go test -bench=. ` 10 times for both branches and fed the input into benchstat. This is from my MBP. It seems to basically be "no change" if I am understanding this correctly?

```
name                                         old time/op    new time/op    delta
Molecule/standard_unmarshal-8                  1.37µs ± 5%    1.39µs ± 9%   ~     (p=0.837 n=10+10)
Molecule/unmarshal_all-8                        162ns ±10%     160ns ± 4%   ~     (p=0.712 n=10+8)
Molecule/unmarshal_loop-8                       150ns ± 7%     154ns ± 5%   ~     (p=0.297 n=9+9)
Molecule/unmarshal_single_with_molecule-8       117ns ±10%     118ns ± 7%   ~     (p=0.374 n=9+9)
Molecule/unmarshal_multiple_with_molecule-8     762ns ± 7%     785ns ±10%   ~     (p=0.400 n=9+10)
Simple-8                                        280ns ± 8%     289ns ±13%   ~     (p=0.368 n=9+10)
Packing-8                                       180µs ±24%     175µs ± 9%   ~     (p=0.912 n=10+10)

name                                         old alloc/op   new alloc/op   delta
Simple-8                                        1.00B ± 0%     1.00B ± 0%   ~     (all equal)
Packing-8                                        813B ± 3%      838B ± 9%   ~     (p=0.139 n=8+9)

name                                         old allocs/op  new allocs/op  delta
Simple-8                                         1.00 ± 0%      1.00 ± 0%   ~     (all equal)
Packing-8                                        0.00           0.00        ~     (all equal)
```